backend/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
email-validator>=2.1.0

# Supabase Integration
supabase==2.0.0
postgrest==0.11.0
gotrue==1.1.0

# Elasticsearch
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0

# Data Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & ORM
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1

# AI & ML Libraries
openai>=1.3.7
anthropic>=0.18.0  # Updated for Agent SDK support
sentence-transformers>=2.2.2
numpy>=1.24.4

# Agent Framework & Evaluation
tiktoken>=0.5.2  # Token counting for agent context management
jsonschema>=4.20.0  # Schema validation for agent tools

# Date & Time
python-dateutil==2.8.2

# Logging & Monitoring
structlog==23.2.0
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Performance & Caching
redis>=5.0.1
# python-redis package doesn't exist - use redis instead
cachetools>=5.3.2

# Utilities
click==8.1.7
rich==13.7.0
typer==0.9.0

# CORS
fastapi-cors==0.0.6

# Rate Limiting
slowapi==0.1.9

# Background Tasks
celery==5.3.4
kombu==5.3.4

# File Processing
PyPDF2==3.0.1
python-magic==0.4.27

# Async Support
# asyncio is built-in, no separate package needed
aiofiles==23.2.1
aiocache==0.12.2

# Web Scraping & HTML Parsing (for enrichment agent)
beautifulsoup4>=4.12.0
lxml>=4.9.0
playwright>=1.40.0

# pSEO Web Scraping (NEW)
firecrawl-py>=0.0.8
aiohttp>=3.9.0
feedparser>=6.0.11

# pSEO Utilities (NEW)
python-slugify>=8.0.0
markdownify>=0.11.6

# Executive Summary Service (NEW)
# Note: Executive summary service uses existing dependencies above
# No additional packages required - leverages FastAPI, Pydantic, etc.